fixed(#all): fix bug of axios request

+++ easyshop-server-commodity/src/main/java/com/github/mayoi7/easyshop/server/service/impl/CartServiceImpl.java
@@ -38,7 +38,7 @@ public class CartServiceImpl implements CartService {
 
     @Override
     public boolean addCart(String username, Long commodityId) {
-        if (redisService.checkExistenceInSet(RedisKeys.CART_LIST, username, commodityId)) {
+        if (redisService.checkExistenceInSet(RedisKeys.CART_LIST, username, commodityId.intValue())) {
             return false;
         }
         redisService.setInSet(RedisKeys.CART_LIST, username, commodityId);
